нужно добавить красивый компонент выбора процента заполненности
