+++ src/App.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import ControlComponent from './gameOfLife2/ControlComponent';
 import { PageText } from './Tests/PageText';
@@ -8,7 +9,7 @@ import { TextStyled } from './Tests/styles';
 function App() {
   return (
     <div className="App">
-      <ControlComponent/>
+      <ControlComponent />
       {/*<TextStyled fontSize={25} color={'#CD5C5C'}/>*/}
     </div>
   );
